back/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
django-allauth==0.63.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
idna==3.7
PyJWT==2.8.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
Pillow==10.4.0